pyproject.toml
[project]
name = "ioc_collector"
version = "0.1.7"
description = "Collect and structure IoC information from security incident reports using Gemini."
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.0.0",
    "pydantic>=2.12.5",
    "stix2>=3.0.1",
    "typer>=0.12.3",
]

[project.urls]
Repository = "https://github.com/nlink-jp/ioc-collector"

[project.scripts]
ioc-collector = "ioc_collector.main:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.3.2",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
]
[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src/ioc_collector"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
